Competency Level Suggestions
Added AI suggestions for competency levels based on the title and description of a course to the admin mask. A level prediction is generated by sending a GET request to admin/wisyki-predict-comp-level.inc.php. The AI model ca be trained by calling admin/wisyki-tryin-comp-level-level.inc.php. In a future version this could be done in a cron-task. The model is trained on a set of prelabeled courses saved as a json file at admin/lib/ki/python/data/labeledCourses.json. These courses are updated with labeled courses from the database, to allow for the model to learn and be trained on newly labeled courses and learn as the platform grows.