Additional CSS and a logo for the scout are set in portal settings
All colors in the main CSS file for the scout are now provided by CSS custom properties. This allows for an easy override of the scouts color scheme other CSS files. Other CSS files can be loaded by setting the property "scout.css" to one or many file paths that are separated by ",". Also a custom logo for the header can be set by providing a file location for the property "scout.logo" in Portale > Einstellungen