Fix visual issues for mobile devices
Detect opening and closing of virtual keyboard on mobile devices to dynamically move away header and footer, to provide more screen space when the keyboard is active. Also added workaround for 'vh', that is not working on mobile devices, by setting a css property with the document height with js.