Update: Several new features and updates
1. New OpenThesaurus and tag statistics tables for search and tag usage
statistics.
[admin/config/db.sql]

2. New CMS feedback plugin which cleans feedback entries:
a) caused by security scans
b) that older than 14 months (previously 12 months)
[admin/config/index_plugin_feedback_0.inc.php]

3. New CMS feedback plugin: list most liked and disliked urls by
visitors (according to use of feedback feature)
[admin/config/index_plugin_feedback_1.inc.php]

4. Updated CMS oververview tables/search result tables, to display
"k.A." when a field is empty, not zero. E.g. relevant for price. Added
new user setting "index.showempty" allows to switch between displaying
nothing or "k.A.".
[admin/index.php, admin/config/lang/settings/de.php, admin/settings.php,
admin/skins/default/cssscreen.php]

5. Changed filter form type in synchronisation settings from text to
textarea for longer/more detailed filter definitions.
[admin/lib/sync/editjobrenderer.inc.php]

6. Minor bug fix (prev./next entry link) for PHP 8.1
[admin/lib/edit/renderer.inc.php]

7. Reactivated login limit and adapted code to consider each logged in
user separately etc.
[admin/login.inc.php]

8. Reduced the output on SQL errors per default for better security.
[admin/sql_curr.inc.php]

9. Added information about licenses of external libraries, content and
files that are included in this repository or are not included but
useful to download and combine in a WISY instance.
[LICENCES_EXTERNAL_LIBRARIES.md]