Added "Merklisten"-feature, added location filter
For the "Merklisten"-feature, all the default js and css had to be imported to enable this feature and all the other features this one depends on, like the cookie banner.
On the text system an error occurs during search when a location filter is active. There seems to be an error with the mapquest api.