Search for similar skills, sort results by most skill-goals met
When a skill is selected in the scout by a user related skills are stored aswell. These similar skills will allow to expand the search capabilities to include courses that teach more specialized skills or skills that are essential before aquiring the target skill.
The search queries are extended to search for mutliple skills at a time and order the results by how many of the target skills a course teaches.
The 5 AI suggestions are composed of the top 3 skill that meet the most target skills and 2 more skills that have the biggest semantic similarity to the user goals. To generate these 2 skills only the next 10 skills after the top 3 skills are sorted by semantic similarity for performance reasons.
WISYKI_SCOUT_SEARCH_CLASS now extends WISY_SEARCH_CLASS directly, since the fulltext search capabilities of WISY_INTELLISEARCH_CLASS are not needed.
The scout now translates the complevels A, B, C to the the respective strings  set in the language file.
The scout now shows a short info message for empty result lists.
For result lists that are filtered by level goal the comlevel is not shown anymore in the course previews. The complevel is shown only for the course previews of the ai-recommendations.
An error is fixed for skills that exist as esco skill as well as wisy stichwort and are therfore marked with the " (ESCO)" in the skill label. This marker is not displayed to the user but now still is kept under the hood, so that the right tag is used in the search queries.