SEO: canonical URL points from sub-portal to main portal if needed
wisy-glossar-renderer-class.inc.php:
- SEO: Make sure html protocol is set correctly for canonical URL and
full url incl. domain is used
- SEO: New portal settings 'seo.canonical.portalid' and
'seo.canonical.domain' allow to point canonical URL of "Glossar" to the
same "Glossar" page in authoritative main portal. This signals Google
that no double content is intended, even if the content of this
sub-portal page is the same as the content of the corresponding URL in
the main portal.